orchestrator/core/requirements.txt
bleach
django
django-cors-headers
django-dynamic-preferences
djangorestframework
djangorestframework-datatables
djangorestframework-jwt
djangorestframework-queryfields
django-rest-swagger
drf-writable-nested
jsonfield
mysql-connector-python
uwsgi